requirements.txt
absl-py==0.7.1
astor==0.8.0
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cmake==3.14.4
comtypes==1.1.7
Cython==0.29.13
Django==2.2.1
firebase-admin==2.17.0
Flask==1.1.1
Flask-Limiter==1.0.1
gast==0.2.2
google-api-core==1.13.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.2
google-cloud-firestore==1.2.0
google-cloud-storage==1.16.1
google-pasta==0.1.7
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.21.1
h5py==2.9.0
httplib2==0.13.0
idna==2.8
image==1.5.27
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
limits==1.3
Markdown==2.6.9
MarkupSafe==1.1.1
memoization==0.1.4
msgpack==0.6.1
mysql-connector==2.2.9
mysqlclient==1.4.2.post1
myutils==0.0.21
numpy==1.16.4
oauth==1.0.1
oauthlib==3.0.2
opencv-contrib-python==3.4.5.20
opencv-python==3.4.5.20
pandas==0.24.2
peewee==3.10.0
Pillow==6.0.0
playsound==1.2.2
protobuf==3.8.0
psycopg2==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyAutoGUI==0.9.45
pycryptodome==3.8.2
PyGetWindow==0.0.6
Pygments==2.4.2
PyMsgBox==1.0.7
PyMySQL==0.9.3
PyPDF2==1.26.0
PyQt5==5.13.0
PyQt5-sip==4.19.18
PyQtWebEngine==5.13.0
PyRect==0.1.4
PyScreeze==0.1.22
PySide2==5.13.0
pytesseract==0.2.7
python-dateutil==2.8.0
PyTweening==1.0.3
pytz==2019.1
requests==2.22.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
shiboken2==5.13.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.8
sqlcrush==0.1.5
sqlite-web==0.3.6
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tesseract==0.1.3
tflearn==0.3.2
tinydb==3.13.0
tqdm==4.32.2
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
wxPython==4.0.6
xlrd==1.2.0
XlsxWriter==1.1.8
xlutils==2.0.0
xlwt==1.3.0
zipfile36==0.1.3
